refactor: move agent creativity in types
