+++ front/components/agent_builder/transformAgentConfiguration.ts
@@ -2,7 +2,6 @@ import uniqueId from "lodash/uniqueId";
 
 import type { AgentBuilderFormData } from "@app/components/agent_builder/AgentBuilderFormContext";
 import type { AgentBuilderMCPConfiguration } from "@app/components/agent_builder/types";
-import { AGENT_CREATIVITY_LEVEL_TEMPERATURES } from "@app/components/agent_builder/types";
 import type { FetchAgentTemplateResponse } from "@app/pages/api/templates/[tId]";
 import type {
   LightAgentConfigurationType,
@@ -10,6 +9,7 @@ import type {
   WorkspaceType,
 } from "@app/types";
 import {
+  AGENT_CREATIVITY_LEVEL_TEMPERATURES,
   CLAUDE_4_5_SONNET_DEFAULT_MODEL_CONFIG,
   getLargeWhitelistedModel,
   isProviderWhitelisted,
