+++ front/components/agent_builder/types.ts
@@ -26,29 +26,12 @@ type AgentBuilderFormData = z.infer<typeof agentBuilderFormSchema>;
 
 export type AgentBuilderAction = AgentBuilderFormData["actions"][number];
 
-export const AGENT_CREATIVITY_LEVELS = [
-  "deterministic",
-  "factual",
-  "balanced",
-  "creative",
-] as const;
-export type AgentCreativityLevel = (typeof AGENT_CREATIVITY_LEVELS)[number];
-
 export const AGENT_CREATIVITY_LEVEL_DISPLAY_NAMES = {
   deterministic: "Deterministic",
   factual: "Factual",
   balanced: "Balanced",
   creative: "Creative",
 } as const;
-export const AGENT_CREATIVITY_LEVEL_TEMPERATURES: Record<
-  AgentCreativityLevel,
-  number
-> = {
-  deterministic: 0.0,
-  factual: 0.2,
-  balanced: 0.7,
-  creative: 1.0,
-};
 
 export const BUILDER_FLOWS = [
   "workspace_assistants",
